scrapper/requirements.txt
google-cloud-bigquery
db-dtypes
lxml
